+++ js/util.js
@@ -1,0 +1,25 @@
+// Функция, возвращающая случайное целое число
+// из переданного диапазона включительно:
+const getRandomPositiveInteger = (a, b) => {
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// Функция, возвращающая случайное число с плавающей
+// точкой из переданного диапазона включительно
+const getRandomPositiveFloat = (a, b, digits = 1) => {
+  if (a < 0 || b < 0 || digits < 0) {
+    return NaN;
+  }
+  const lower = Math.min(a, b);
+  const upper = Math.max(a, b);
+  const result = Math.random() * (upper - lower) + lower;
+  return +result.toFixed(digits);
+};
+
+export {getRandomPositiveInteger, getRandomPositiveFloat};
